PettingMain class added
This class is for Petting Module only. Responsible for all the game play inside the Petting Game. Restrictions, gestures, transitions etc. This handles it all and this is one of the biggest module on the game since this is where the main core loop is. The Dragon which is your pet is driven by data input created in GoogleDocs Excel and parsed using Ruby script to generate dragon's behavioural state machine.